Add an integration test for qanban-node

+++ node/src/index.ts
@@ -254,7 +254,7 @@ if (args.database === undefined) {
 } else {
   database = args.database;
 }
-if (args.clean) {
+if (args.clean && fs.existsSync(database)) {
   fs.unlinkSync(database);
 }
